[sol] 디스크 컨트롤러 kotlin 풀이
